Move ML-Plan model selection to new Maven module
