+++ swan-cmd/pom.xml
@@ -106,21 +106,6 @@
             <version>${dl4j-master.version}</version>
         </dependency>
         <dependency>
-            <groupId>ai.libs</groupId>
-            <artifactId>mlplan-meka</artifactId>
-            <version>0.2.7</version>
-        </dependency>
-        <dependency>
-            <groupId>org.graphstream</groupId>
-            <artifactId>gs-ui-javafx</artifactId>
-            <version>2.0</version>
-        </dependency>
-        <dependency>
-            <groupId>ai.libs</groupId>
-            <artifactId>hasco-core</artifactId>
-            <version>0.2.7</version>
-        </dependency>
-        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
             <version>1.7.36</version>
