+++ swan-cmd/src/main/java/de/fraunhofer/iem/swan/features/FeatureSetSelector.java
@@ -10,7 +10,6 @@ public class FeatureSetSelector {
 
         switch (ModelEvaluator.Toolkit.valueOf(options.getToolkit().toUpperCase())) {
 
-            case ML2PLAN:
             case MEKA:
                 MekaFeatureSet mekaFeatureSet = new MekaFeatureSet(dataset, options);
                 mekaFeatureSet.createFeatures();
